+++ datanode/entities/order_filter.go
@@ -6,5 +6,10 @@ type OrderFilter struct {
 	Statuses         []vega.Order_Status
 	Types            []vega.Order_Type
 	TimeInForces     []vega.Order_TimeInForce
+	Reference        *string
+	DateRange        *DateRange
 	ExcludeLiquidity bool
+	LiveOnly         bool
+	PartyIDs         []string
+	MarketIDs        []string
 }